+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- Fixed an issue where App Store ID was always prompted for even when unnecessary.
+- Fixed an issue in `apps:create` where App Store ID was always prompted for even when unnecessary.
+- Add `functions:lifecycle:list` and `functions:lifecycle:run` commands to view and run
+  lifecycle hooks in isolation.